Move supported operations module from c8y_api to c8y_mapper_ext
`c8y_api::smartrest::operations` was moved to
`c8y_mapper_ext::supported_operations` because:

- it deals with loading operation files from the local filesystem, not
  c8y_api concern
- it was inside `smartrest` module, while not having much to do with
  smartrest (ie. it only generated "114 Set supported operations"
  message)
