+++ crates/core/c8y_api/src/smartrest/mod.rs
@@ -3,7 +3,6 @@ pub mod csv;
 pub mod error;
 pub mod inventory;
 pub mod message;
-pub mod operations;
 pub mod payload;
 pub mod smartrest_deserializer;
 pub mod smartrest_serializer;
