+++ crates/extensions/c8y_mapper_ext/src/converter.rs
@@ -11,6 +11,12 @@ use crate::error::MessageConversionError;
 use crate::json;
 use crate::operations;
 use crate::operations::OperationHandler;
+use crate::supported_operations::get_child_ops;
+use crate::supported_operations::Operation;
+use crate::supported_operations::Operations;
+use crate::supported_operations::OperationsError;
+use crate::supported_operations::ResultFormat;
+use crate::supported_operations::SupportedOperations;
 use anyhow::anyhow;
 use anyhow::Context;
 use c8y_api::json_c8y::C8yCreateEvent;
@@ -19,7 +25,6 @@ use c8y_api::json_c8y_deserializer::C8yDeviceControlTopic;
 use c8y_api::json_c8y_deserializer::C8yJsonOverMqttDeserializerError;
 use c8y_api::json_c8y_deserializer::C8yOperation;
 use c8y_api::json_c8y_deserializer::C8ySoftwareUpdate;
-use c8y_api::smartrest::error::OperationsError;
 use c8y_api::smartrest::error::SmartRestDeserializerError;
 use c8y_api::smartrest::inventory::child_device_creation_message;
 use c8y_api::smartrest::inventory::service_creation_message;
@@ -29,11 +34,6 @@ use c8y_api::smartrest::message::get_smartrest_device_id;
 use c8y_api::smartrest::message::get_smartrest_template_id;
 use c8y_api::smartrest::message::sanitize_bytes_for_smartrest;
 use c8y_api::smartrest::message::MAX_PAYLOAD_LIMIT_IN_BYTES;
-use c8y_api::smartrest::operations::get_child_ops;
-use c8y_api::smartrest::operations::Operation;
-use c8y_api::smartrest::operations::Operations;
-use c8y_api::smartrest::operations::ResultFormat;
-use c8y_api::smartrest::operations::SupportedOperations;
 use c8y_api::smartrest::smartrest_serializer::fail_operation_with_id;
 use c8y_api::smartrest::smartrest_serializer::fail_operation_with_name;
 use c8y_api::smartrest::smartrest_serializer::request_pending_operations;
@@ -1674,12 +1674,12 @@ pub(crate) mod tests {
     use crate::actor::IdUploadResult;
     use crate::config::BridgeConfig;
     use crate::config::C8yMapperConfig;
+    use crate::supported_operations::ResultFormat;
     use crate::Capabilities;
     use anyhow::Result;
     use assert_json_diff::assert_json_include;
     use c8y_api::json_c8y_deserializer::C8yDeviceControlTopic;
     use c8y_api::proxy_url::Protocol;
-    use c8y_api::smartrest::operations::ResultFormat;
     use c8y_api::smartrest::topic::C8yTopic;
     use c8y_http_proxy::handle::C8YHttpProxy;
     use serde_json::json;
