+++ crates/extensions/c8y_mapper_ext/src/dynamic_discovery.rs
@@ -1,4 +1,4 @@
-use c8y_api::smartrest::operations::is_valid_operation_name;
+use crate::supported_operations::is_valid_operation_name;
 use serde::Deserialize;
 use serde::Serialize;
 use std::path::Path;
