+++ crates/extensions/c8y_mapper_ext/src/lib.rs
@@ -12,6 +12,8 @@ pub mod json;
 mod operations;
 mod serializer;
 pub mod service_monitor;
+mod supported_operations;
+
 #[cfg(test)]
 mod tests;
